Add files via upload
Splitting the file
