+++ desiredAverageCalculate.h
@@ -1,0 +1,8 @@
+#ifndef DESIREDAVERAGECALCULATE_H
+#define DESIREDAVERAGECALCULATE_H
+
+int scoreNeededForAverage(int takenSoFar, int currentAverage, int taking, int wantedAverage);
+
+void desiredAverage();
+
+#endif
